LPS-61952 Move com.liferay.portlet.admin.util to com.liferay.admin.kernel.util
