+++ portal-service/src/com/liferay/admin/kernel/util/PortalDirectoryApplicationType.java
@@ -12,18 +12,18 @@
  * details.
  */
 
-package com.liferay.portlet.admin.util;
+package com.liferay.admin.kernel.util;
 
 /**
- * @author Eudaldo Alonso
+ * @author Pei-Jung Lan
  */
-public class PortalProductMenuApplicationType {
+public class PortalDirectoryApplicationType {
 
-	public interface ProductMenu {
+	public interface PortalDirectory {
 
 		public static final String CLASS_NAME =
-			"com.liferay.portlet.admin.util." +
-				"PortalProductMenuApplicationType$ProductMenu";
+			"com.liferay.admin.kernel.util." +
+				"PortalDirectoryApplicationType$PortalDirectory";
 
 	}
